pyproject.toml
[project]
name = "demo-chat-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.25",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "python-dotenv>=1.1.0",
]
